# shared-contracts/openapi/Sacco.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SACCO Platform API",
    "description": "Core banking API for Kenyan Deposit-Taking SACCOs. Pass the tenant slug in the X-Tenant header when not resolving by host. Tokens are issued by the in-process OIDC server at /connect/*.",
    "version": "v1"
  },
  "paths": {
...
    "/api/loans/products": {
      "get": {
        "tags": [
          "Loan products"
        ],
        "operationId": "ListLoanProducts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanProductResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Loan products"
        ],
        "operationId": "CreateLoanProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoanProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/products/{code}": {
      "put": {
        "tags": [
          "Loan products"
        ],
        "operationId": "UpdateLoanProduct",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLoanProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "ListLoans",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/LoanStatus"
            }
          },
          {
            "name": "memberId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfLoanListItem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "ApplyForLoan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyLoanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/eligibility": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "GetLoanEligibility",
        "parameters": [
          {
            "name": "memberId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "productCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibilityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/guarantors/{memberId}/exposure": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "GetGuarantorExposure",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuarantorExposure"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/{id}": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "GetLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/loans/{id}/guarantors": {
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "AddLoanGuarantor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGuarantorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuarantorResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/{id}/guarantors/{guarantorId}/accept": {
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "AcceptLoanGuarantee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "guarantorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuarantorResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/{id}/guarantors/{guarantorId}/decline": {
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "DeclineLoanGuarantee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "guarantorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuarantorResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/{id}/appraise": {
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "AppraiseLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/{id}/approve": {
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "ApproveLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/{id}/reject": {
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "RejectLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddGuarantorRequest": {
        "required": [
          "guarantorMemberId",
          "amount"
        ],
        "type": "object",
        "properties": {
          "guarantorMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "LoanStatus": {
        "enum": [
          "Applied",
          "Appraised",
          "PendingApproval",
          "Approved",
          "Active",
          "Closed",
          "Rejected",
          "Cancelled",
          "WrittenOff"
        ]
      },
...
      "NotesRequest": {
        "required": [
          "notes"
        ],
        "type": "object",
        "properties": {
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedResultOfLoanListItem": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanListItem"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}